+++ src/main.rs
@@ -1,4 +1,5 @@
 mod config;
+mod php;
 mod server;
 mod watcher;